requirements.txt
pyyaml>=4.2b1
requests==2.31.0
baidu-aip==2.0.0.1
pydub==0.23.1
python-dateutil==2.7.5
watchdog==0.9.0
pytz==2018.9
fire==0.1.3
tornado==6.3.2
markdown==3.0.1
semver==2.8.1
websocket==0.2.1
websocket-client
pypinyin
jieba
pvporcupine
pvrecorder
openai
apscheduler
asyncio
edge-tts
nest_asyncio
langchain==0.0.137
pinecone-client==2.2.1